requirements.txt
pandas==2.2.1
GitPython==3.1.42
tk==0.1.0
pyodbc==5.1.0
azure-storage-blob==12.19.1
python-dotenv==1.0.1
azure-identity==1.16.0
SQLAlchemy~=2.0.29